Round 57 wrap-up: two coherence repairs (r57d), receipt, research note, handover and mechanism docs; 5,859 pins in 1,251 rows

+++ src/dungeon/func_808B16A0.c
@@ -12,7 +12,7 @@ typedef struct S_808B16A0_0 {
 /* Pass both signed 16-bit values in each pair to func_80016BDC. */
 void func_808B16A0(void *pairs, s32 pairCount) {
     char pad[8];
-    s16 secondValue;
+    s32 secondValue;
     s32 pairsProcessed;
     void *currentPair;
